separated test groups
separated tests into logical grouping
consolidated imorts
